requirements.txt
appdirs==1.4.4
astor==0.8.1
Babel==2.9.1
bce-python-sdk==0.8.61
cached-property
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
colorlog==5.0.1
cycler==0.10.0
decorator==4.4.2
dill==0.3.4
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
future==0.18.2
gast==0.5.0
h5py
identify==2.2.10
idna==2.10
importlib-metadata==4.6.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random
mkl-service==2.3.0
multiprocess==0.70.12.2
nodeenv==1.6.0
numpy
paddlenlp==2.0.5
paddlepaddle==2.1.1
pandas==1.3.0
Pillow==8.3.0
pre-commit==2.13.0
prettytable==2.1.0
protobuf==3.17.3
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seqeval==1.2.2
shellcheck-py==0.7.2.1
six
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.4.7
visualdl==2.2.0
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
